Allow t-b-m to be restarted if it exits with 2
Although this is a good idea in theory, in practice it can often happen
that a postgresql connection cannot be made at exactly that moment. With
the previous behaviour, the entire t-b-m would stop and not start.

When PG is restarted, then the connections are interrupted and in the
old behaviour, t-b-m would stop & stay stopped.

More common is timeouts. Often t-b-m workers are configured to restart
every X requests (e.g. 1000). Mapnik by default includes a 4 sec timeout
when opening a pg conn. If the load is a little high, this timeout can
be reached, causing the t-b-m to stop and stay stopped. In reality, this
could attempted to be restarted.